Refactored `autopkg/receive`'s event handling into it's own file
