+++ pkgbot/api/__init__.py
@@ -1,9 +1,9 @@
 from . import auth
-from . import autopkg
 from . import package
-from . import recipe 
+from . import recipe
 from . import user
 from . import views
+from .autopkg import autopkg
 from .slack import block_builders
 from .slack import bot
 from .slack import build_msg
