+++ pkgbot/api/autopkg/events.py
@@ -1,0 +1,341 @@
+import time
+
+from datetime import datetime
+from tempfile import SpooledTemporaryFile
+
+from fastapi import UploadFile
+
+from pkgbot import api, config
+from pkgbot.db import models
+from pkgbot.tasks import task_utils
+from pkgbot.utilities import common as utility
+
+
+config = config.load_config()
+log = utility.log
+
+
+async def event_handler(task_id):
+
+	task_results = task_utils.get_task_results(task_id)
+
+	try:
+		event = task_results.result.get("event")
+		# event_id = task_results.result.get("event_id", "")
+		# called_by = task_results.result.get("called_by")
+		# recipe_id = task_results.result.get("recipe_id")
+		# success = task_results.result.get("success")
+		# stdout = task_results.result.get("stdout")
+		# stderr = task_results.result.get("stderr")
+		# callback = await determine_callback(called_by)
+
+	except Exception:
+		# Seeing a random issue where looking up the task_id returns "None" here, but if I manually
+		# lookup the task_id later, the results are found and the task completed successfully...
+		# It's almost always the _very last_ recipe to run as well (after ~100 recipes).
+		# Ideas:
+			# Due to task prioritization?
+				# But...why when the task is essentially finished when the webhook is sent
+				# And it just started (or I just started noticing it)
+		log.error(f"FAILED GETTING TASK RESULTS for {task_id} -- WHY?!?")
+		log.debug(f"task_results:  {task_results}")
+		log.debug(f"task_results.result:  {task_results.result}")
+		log.debug("Sleeping for one second...")
+		time.sleep(1)
+		await event_handler(task_id)
+
+	if event == "verify_trust_info":
+		await event_verify_trust_info(task_results)
+
+	elif event == "update_trust_info":
+		await event_update_trust_info(task_results)
+
+	elif event in {"autopkg_repo_update", "disk_space_critical", "private_git_pull"}:
+		await event_failed_pre_checks(task_results)
+
+	elif event == "disk_space_warning":
+		await event_disk_space_warning(task_results)
+
+	elif event == "failed_pre_checks":
+		await event_failed_pre_checks(task_results)
+
+	elif event == "error" or not task_results.result.get("success"):
+		await event_error(task_results)
+
+	elif event in ("recipe_run_dev", "recipe_run_prod"):
+		await event_recipe_run(task_results)
+
+
+async def event_details(task_results):
+
+	return ( task_results.result.get("event"), #"event":
+		task_results.result.get("event_id", ""), #"event_id":
+		task_results.result.get("called_by"), #"called_by":
+		task_results.result.get("recipe_id"), #"recipe_id":
+		task_results.result.get("success"), #"success":
+		task_results.result.get("stdout"), #"stdout":
+		task_results.result.get("stderr") #"stderr":
+	)
+	# or: {
+		# "event": task_results.result.get("event"),
+		# "event_id": task_results.result.get("event_id", ""),
+		# "called_by": task_results.result.get("called_by"),
+		# "recipe_id": task_results.result.get("recipe_id"),
+		# "success": task_results.result.get("success"),
+		# "stdout": task_results.result.get("stdout"),
+		# "stderr": task_results.result.get("stderr")
+	# }
+
+
+async def event_verify_trust_info(task_results):
+	""" When a user has requested to run `autopkg verify-trust-info <recipe>` """
+
+	event, event_id, called_by, recipe_id, success, stdout, stderr = await event_details(task_results)
+	callback = await determine_callback(called_by)
+
+	if callback in [ "PkgBot", "api" ]:
+
+		if success:
+			# This shouldn't ever be called?
+			log.info(f"Trust info verified for:  {recipe_id}")
+
+		else:
+			# Send message that recipe_id failed verify-trust-info
+			redacted_error = await utility.replace_sensitive_strings(stderr)
+			await api.recipe.recipe_trust_verify_failed(recipe_id, redacted_error)
+
+	elif callback == "ephemeral":
+##### TO DO:
+		log.debug("Recipe trust info was checked via Slack command.")
+		# Post ephemeral msg to Slack user
+
+		if success:
+			# trust info verified msg
+			pass
+		else:
+			# trust info invalid msg
+			pass
+
+
+async def event_update_trust_info(task_results):
+	""" Update message with result of update-trust-info attempt """
+
+	event, event_id, called_by, recipe_id, success, stdout, stderr = await event_details(task_results)
+
+	if success:
+		await api.recipe.recipe_trust_update_success(event_id)
+	else:
+		await api.recipe.recipe_trust_update_failed(event_id, str(stderr))
+
+
+async def event_disk_space_warning(task_results):
+	""" If cache volume has low disk space """
+
+	event, event_id, called_by, recipe_id, success, stdout, stderr = await event_details(task_results)
+
+	await api.send_msg.disk_space_msg("Warning", stderr, config.PkgBot.get('icon_warning'))
+
+
+async def event_failed_pre_checks(task_results):
+	""" When a pre-check task fails """
+
+	event, event_id, called_by, recipe_id, success, stdout, stderr = await event_details(task_results)
+
+	for task_id in task_results.result.get("task_id"):
+
+		task_results = task_utils.get_task_results(task_id)
+		event = task_results.result.get("event")
+
+		if event == "autopkg_repo_update":
+##### TODO:
+			pass
+
+		if event == "disk_space_critical":
+			""" If cache volume has insufficient disk space """
+
+			await api.send_msg.disk_space_msg(
+				"Critical",
+				task_results.result.get("stderr"),
+				config.PkgBot.get('icon_error')
+			)
+
+		if event == "private_git_pull":
+##### TODO:
+			pass
+
+
+async def event_error(task_results):
+	""" When a recipe run or unknown event fails """
+
+	event, event_id, called_by, recipe_id, success, stdout, stderr = await event_details(task_results)
+
+	await handle_autopkg_error(task_id = task_results, event = event, event_id = event_id,
+		called_by = called_by, recipe_id = recipe_id, success = success, stdout = stdout,
+		stderr = stderr
+	)
+
+
+async def event_recipe_run(task_results):
+
+	event, event_id, called_by, recipe_id, success, stdout, stderr = await event_details(task_results)
+	callback = await determine_callback(called_by)
+	plist_contents = await utility.find_receipt_plist(stdout)
+
+	# Get the log info for PackageUploader
+	pkg_processor = await utility.parse_recipe_receipt(
+		plist_contents, "JamfPackageUploader")
+	pkg_name = pkg_processor.get("Output").get("pkg_name")
+	pkg_data = {
+		"name": (pkg_name).rsplit("-", 1)[0],
+		"pkg_name": pkg_name,
+		"recipe_id": recipe_id,
+		"version": pkg_processor.get("Input").get("version"),
+		"notes": pkg_processor.get("Input").get("pkg_notes")
+	}
+
+	if event == "recipe_run_dev":
+
+		try:
+			# Get the log info for PolicyUploader
+			policy_processor = await utility.parse_recipe_receipt(
+				plist_contents, "JamfPolicyUploader")
+			policy_results = policy_processor.get(
+				"Output").get("jamfpolicyuploader_summary_result").get("data")
+			pkg_data["icon"] = policy_results.get("icon")
+
+			# Create a temporary file to hold the icon data and upload it.
+			# This is required since we're not actually using an
+			# HTTP client to interface with the API endpoint.
+			icon_data = SpooledTemporaryFile()
+			with open(policy_results.get("icon_path"), "rb") as icon_path:
+				icon_data.write(icon_path.read())
+			_ = icon_data.seek(0)
+			icon = UploadFile(filename=pkg_data["icon"], file=icon_data)
+			await api.views.upload_icon(icon)
+
+		except Exception:
+			log.info(
+				f"An icon was not identified, therefore it was not uploaded into PkgBot.  Review task_id:  {task_results}")
+
+		try:
+
+			# No, don't check the processor summary...
+			# if pkg_processor.get("Output").get("pkg_uploaded"):
+
+			# Instead, check if the package has already been created in the database, this
+			# ensures a message is posted if it failed to post previously.
+			pkg_db_object = await models.Packages.filter(pkg_name=pkg_name).first()
+
+			if pkg_db_object and callback == "ephemeral":
+##### TO DO:
+# Post ephemeral msg to Slack user
+					log.debug("Post ephemeral msg to Slack user new software version was not available")
+
+			elif not pkg_db_object:
+				log.info(f"New package posted to dev:  {pkg_name}")
+				await api.autopkg.workflow_dev(models.Package_In(**pkg_data))
+
+				if callback == "ephemeral":
+##### TO DO:
+# Post ephemeral msg to Slack user
+					log.debug("Post ephemeral msg to Slack user that a new pkg was posted")
+
+			# Update attributes for this recipe
+			recipe_object = await models.Recipes.filter(recipe_id=recipe_id).first()
+			recipe_object.last_ran = await utility.utc_to_local(datetime.now())
+			recipe_object.recurring_fail_count = 0
+			await recipe_object.save()
+
+		except Exception as exception:
+
+			await handle_exception(task_id = task_results, event = event, event_id = event_id,
+				called_by = called_by, recipe_id = recipe_id, success = success,
+				exception = exception
+			)
+
+	elif event == "recipe_run_prod":
+		log.info(f"Package promoted to production:  {pkg_name}")
+
+		format_string = "%Y-%m-%d %H:%M:%S.%f"
+		promoted_date = datetime.strftime(datetime.now(), format_string)
+		pkg_data["promoted_date"] = promoted_date
+
+		await api.autopkg.workflow_prod(event_id, models.Package_In(**pkg_data))
+
+
+async def determine_callback(caller: str):
+
+	if caller == "schedule":
+		return "PkgBot"
+
+	if caller == "slack":
+		return "ephemeral"
+
+	if caller == "api":
+		return "api"
+
+
+async def handle_autopkg_error(**kwargs):
+
+	task_id = kwargs.get("task_id")
+	event = kwargs.get("event")
+	event_id = kwargs.get("event_id")
+	called_by = kwargs.get("called_by")
+	recipe_id = kwargs.get("recipe_id")
+	success = kwargs.get("success")
+	stdout = kwargs.get("stdout")
+	stderr = kwargs.get("stderr")
+
+	# Post message with results
+	log.error(f"Failed running:  {recipe_id}")
+
+	try:
+		plist_contents = await utility.find_receipt_plist(stdout)
+		run_error = await utility.parse_recipe_receipt(plist_contents, "RecipeError")
+	except Exception:
+		run_error = stderr
+
+	redacted_error = await utility.replace_sensitive_strings(run_error)
+
+	if event == "recipe_run_prod":
+		# Promotion Failed
+##### Possible ideas:
+# Thread the error message with the original message?
+# Post Ephemeral Message to PkgBot Admin?
+
+		# Get the recipe that failed to be promoted
+		pkg_db_object = await models.Packages.filter(id=event_id).first()
+		recipe_id = pkg_db_object.recipe_id
+		software_title = pkg_db_object.name
+		software_version = pkg_db_object.version
+		log.error(f"Failed to promote:  {pkg_db_object.pkg_name}")
+
+		redacted_error = {
+			"Failed to promote:": f"{software_title} v{software_version}",
+			"Error:": redacted_error
+		}
+
+	await api.recipe.recipe_error(recipe_id, redacted_error, task_id)
+
+
+async def handle_exception(**kwargs):
+
+	task_id = kwargs.pop("task_id")
+	recipe_id = kwargs.pop("recipe_id")
+	event = kwargs.get("event")
+	event_id = kwargs.get("event_id")
+	called_by = kwargs.get("called_by")
+	success = kwargs.get("success")
+	exception = await utility.replace_sensitive_strings(kwargs.get("exception"))
+
+	redacted_error = {
+		"Encountered Exception:": {
+			"Event:": event,
+			"Event ID:": event_id,
+			"Success:": success,
+			# "Called By:": called_by,
+			"Exception:": exception
+		}
+	}
+
+	await api.recipe.recipe_error(recipe_id, redacted_error, task_id)
